+++ main1033.cpp
@@ -14,7 +14,7 @@ int main() {
  unsigned long long Jco78oJY1 = 2061335904;
  float c713 = 664192196;
 	char	Pp5iZav2	=	-1423278364;
- short int FOd0Tiz = -1269412835;
+ short int FOd0Tiz = (short)-1269412835;
 	short int	gIiG9	=	-1239327019;
     float    kWyr5    =    -1097125019;
 	char	W6VQ42G46N	=	1126766103;
